+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   KeySerialization, SignatureSerialization}`.
 - Improved hex formatting of `types::Id`:
   - Removed the unused `Id::hex`.
+  - Deprecated `Id::hex_path` and added `Id::legacy_hex_path` as a replacement.
 
 ### Fixed